+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.2] - 22 FEB 21
+
+ * Prepared for pub.dev
+    * Updated comments and documentation; fixed M⬇; ran `dartdoc`.
+    * Cleaned and linked READMEs; added hyperlinks to Surface Example.
+* Removed Color and Duration constants
+* Goodie [scaleAxis] simplified
+
 ## [0.0.1] - 22 FEB 21
 
  * Work in progress transfer from an [Android application](https://play.google.com/store/apps/details?id=com.zaba.bug_bash 'Bug Bash in the Play Store') on which I am primarily focused.